+++ blocklib/fileio/file_sink/file_sink_cpu.cc
@@ -34,7 +34,7 @@ work_return_code_t file_sink_cpu::work(std::vector<block_work_input_sptr>& work_
         d_itemsize = work_input[0]->buffer->item_size();
     }
 
-    int nwritten = 0;
+    size_t nwritten = 0;
 
     do_update(); // update d_fp is reqd
